bcmdhd: Fixed to handle sets/tags of the event log properly in driver
Basically, the event log is composed of a combination of TAGs and SETs.
However, the driver doesn't handle its SETs properly. (settings mismatch between DHD/FW)
In the current implementation of the driver,
the SET of all event logs are associated with 'SET:0' and 'SET:1' by default.
Hence, it leading to unwanted wake-up frequently.
This is an obvious bug and needs to be fixed properly.

Bug: 180538011
Test: Confirmed by @ramanseth, no longer periodic/frequent host wakeups are observed
Change-Id: Ib96b5b92ba94900b451295df441d6b346925be85
Signed-off-by: Roger Wang <wangroger@google.com>