Firmware roaming not getting disabled
The framework and dhd roam enable/disable values are out of sync. Also,
DHD enables roam back when there is only one STA connected and that
will cause the roam to be enabled back even if framework has instructed
to keep the roam disabled. Added a variable to store the framework roam
enable/disable state based on which DHD will configure roam.

Bug:187904990

Signed-off-by: Jithu Jance <jithu.jance@broadcom.corp-partner.google.com>
Change-Id: I255bf8b38cab86f80ded958190029e8ca57ec49a