bcmdhd: Fixed to increase on_time_scan of link stats monotonically
scan time is not monotonically increasing when we switch 5G->2G->5G since BRCM is using a value per core, and not accumulative.
Logs in Refer https://b.corp.google.com/issues/185731044#comment1
As discussed in the weekly call, we fixed it by adding a cache mechanism as WAR.
The purpose is to give a monotonical value regardless of the connected core, even if it is not accurate.
We have planned to handle all of them in the next project in the proper way.

Bug: 181271600
Test: on_time_scan value (as well as on_time_pno_scan and on_time_roam_scan) increase monotonically even if STA associated from 5G to 2G.
Change-Id: Idee36817ef738d2dacb0d8d698cea73a14bbd3b2