Fix for scan timeout caused by regular scan-pno race condition
Due to race condition with PNO and regular scan, the abort context
didn't clear the scan timer. The two instances of scan triggered at
same time and the scan_request data structure of the on-going scan
got overwritten by the new scan context. This caused a mismatch when
notify_escan_complete got invoked and resulted in early exit.

Bug:187076169
Bug:187602229
Test: Verify issue is not reproducible

Signed-off-by: Jithu Jance <jithu.jance@broadcom.corp-partner.google.com>
Change-Id: I6a7f02c1c145cea69bb10d7a547072b7e772788f