bcmdhd: Fixed to get mgmt tx packet fate properly
Currently, the driver uses 'status_pos' instead of 'pkt_pos' to determine tx pkt_count. (In dhd_dbg_monitor_get_tx_pkts)
That is, pkt_count is to be 0 if there is no data packet. (it means no txstatus update)
This is a bug since mgmt pkt logging feature implemented during P21 project.
=> If without mgmt packet logging, the using 'status_pos' is right way.
If mgmt logger is used, it should be modified to use 'pkt_pos'.

Bug: 190588448
Test: TX/RX packets fates including mgmt packets sends-up properly
Change-Id: I913e6044a6e85765d15e5aa9c73204b8205a7f5e
Signed-off-by: Roger Wang <wangroger@google.com>