bcmdhd: Fixed to set 'dtim multiplier' properly
Pixel 6 has close to 4x 'radio on time' at 2G at high RSSI.
On the other hand, these two devices have similar 'radio on time' at 5G.

According to the log, 'bcn_li_dtim' set to ‘1’ even though LCD-off suspend.
It should be set to '2' because 'dtim multiplier' set to 2 under IPv6 network.

Eventually, we found a bug for 'dtim multiplier' logic.
About few months ago, we refactored the related code, there was a mistake.
The type of 'suspend_bcn_li_dtim' variable changed to boolean,
so that bcn_li_dtim is set to 1 always.
So, fixed it to integer as we intended.

The customer confirmed that bcn_li_dtim is a dominant factor in this issue.
If it set to 3 (with DTIM 3), it seems significantly improved.

Bug: 193811702
Test: Kai confirmed that 'bcn_li_dtim' is a dominant factor, it seems significantly improved if it set to 3 under DTIM 3.
Change-Id: I04f4e640485314773a1e08a5b5dc37c8d6494545