bcmdhd: Fixed S2MPU fault due to wrong DMA direction
This bug is forked from b/193978926,
after S2MPU code update, the S2MPU fault detected as below :

07-18 00:39:23.279 root 0 0 E : [ C3] s2mpu 14480000.s2mpu: IRQ (254): FAULT_STATUS=0x2
07-18 00:39:23.279 root 0 0 E : [ C3] PA=0x0x000000093f76e0a0
07-18 00:39:23.279 root 0 0 E : [ C3] DIRECTION=read
07-18 00:39:23.279 root 0 0 E : [ C3] FAULT_TYPE=[n/a]
07-18 00:39:23.279 root 0 0 E : [ C3] FAULT_INFO=0x1450000

By checking the dump, we could find the erroneous host address in Rx DMA.
To fix the problem, we change Rxbufferpost DMA direction to BIDIRECTIONAL.

The reason of the bidirectional operation is :
there are certain Rx data frames which are processed in the dongle (not to be consumed by the host. For example TDLS, Neighbour Discovery(IPV6) offload, etc),
since these are data packets, they will be received in the host rx buffer by default. As part of processing these frames in firmware, dongle identifies them, schedules pktfetch & reads the entire host rx buffer to fetch the payload to dongle for processing them further.

Bug: 194239976
Bug: 193978926
Test: Run and pass traffic, SSR issue is no loner reproducible

Signed-off-by: Ahmed ElArabawy <arabawy@google.com>
Change-Id: I05152f578d9975eaaea420e2bb9e5906bff542b4