bcmdhd: Use compile time macro for enabling disabling RFC8325 support
The support for 8325 has caused WMM-AC 5.1.X test case failures. so this patch adds a compile time macro for the change so that it can be enabled as per customer requirements.
We found a possible cause of this test case failure. There is a logic that maps UP6 priority to UP7 to support RFC8325 recommendation.
This has been requested and used by other customers. We shall share a CL to confirm this indeed is the problem.

++++++ From RFC RFC8325
However, by default, packets marked
DSCP CS6 will be mapped to UP 6 and serviced within the Voice Access
Category (AC_VO); such mapping and servicing is a contradiction to
the intent expressed in section 3.2 As such, it is
RECOMMENDED to map Network Control Traffic marked CS6 to UP 7 (per
[I80211-2016] Section 10.2.4.2, Table 10-1), thereby admitting
it to the Voice Access Category (AC_VO), albeit with a marking
distinguishing it from (data-plane) voice traffic.

Bug: 192557025
Test: Validated by Paul and BRCM cert team
Change-Id: Ic23440553f45a2ecc2442a91a6acfd581c48dac8