bcmdhd: Fixed to clear EXCESS_PM_PERIOD explicitly when Wi-Fi turn off
DPM(MPC_DUR_EXCEEDED) alert event was raised when WiFi was turned off for a long time
and left for a long time and then turned on again. Generally, WiFi OFF/ON is done through the UI,
so the EXCESS_PM_PERIOD setting is initialized because of LCD ON.
Hence, it is a problem that will not happen in end-user environment,
and it will not have any impact because it is not connected in the latest driver. (pa/1976086)

To avoid this scenario,
we can clear excess_pm_period settings explicitly when WiFi turn off even though LCD OFF state.

Bug: 196011742
Test: settings is getting cleared when Wi-Fi turn off regardless of LCD ON/OFF state
Change-Id: I5155962b6118698c6cee7791204819dc5271b6d0
Signed-off-by: Roger Wang <wangroger@google.com>