bcmdhd: Fixed to check device for debug ring
According to the dump, the wdev (ieee8021_ptr) was null at wl_cfgvendor_dbg_ring_send_evt().
It seems that after triggering a dongle trap, the net device was down but the pointer of ndev was used for a debug ring process.
When ndev was released, then it release wdev (ieee8021_ptr) and set it to NULL also,
so we need to check the validity also of ndev with ieee8021_ptr to protect accessing unregisted device.

Bug: 195972245
Test: MTBF completed
Change-Id: I803c7015546e72168645b9ba92199d90bf1d9c1b
Signed-off-by: Roger Wang <wangroger@google.com>