bcmdhd: Fixed ACS request setting with bw80
The issue is the low tput on soft ap because it was raised on channel 165 with bandwidth 20Mhz even though the requested bandwidth is 80.
The analysis and solution is :

1. When driver makes the chanspec list from the frequency list, only the bw20 chanspecs are included in the result.
The reason is there are some definitions with unsigned symbol 'u' after the number, it breaks the chanspec loop in case of bw 40,80 and 160.

2. FW supports maximum 32 chanspecs for the APCS request, so we don't include all sub bandwidths for the safety margin.

3. ch165 has bw20 only so we must filter the channel out for the other bandwidths.
If chanspecs are set with others (165/40, 165/80, 165/160), then ACPS would return an error.

Bug: 196510315
Test: Validated with Hikey
Test: Verified on latest version and hotspot is able to raise on 80Mhz
Change-Id: I6976b91906362f75fd717497a2820446f2f26229
Merged-In: I6976b91906362f75fd717497a2820446f2f26229

Signed-off-by: Roger Wang <wangroger@google.com>
(cherry picked from commit b64af32676698abcafddd07305023840a626d18d)

Change-Id: I31c6b6be4dcae51f2e107c40eeb291b6fb29e733