bcmdhd: Added debug log for channel avoidance feature
Occasionally, Hotspot turning on failed due to ACS logic.
In that case, target chanspec is not found in avail list/cellular list.
So, it resulting in phone stuck and reboot is required to recover.
These debug logs will tell us why/how chanspec is removed in avail list/cellular list.

Bug: 199128726
Test: build done, it works properly on Hikey board.
Change-Id: Ia8efc6a9d48d438a0b19d7f4aa559c9564490102
Signed-off-by: Roger Wang <wangroger@google.com>