bcmdhd: Fixed ACS failure on 2G with RSDB
If STA connected to AP with DFS channel on 5G, then the control channel of DFS is not listed in the safe channel list. So SoftAP should use the RSDB scenario instead of SCC, i.e. SoftAP uses 2G band.
There is no parts to adding 2G safe channels in parsing ACS parameters and sometimes the channel or freq list from host was not set properly, it leads to 0 candidate for ACS request to dongle.
So we fix to save the candidate chanspec for 2G channels from the safe channel list.

Bug: 200126921
Test: basic test passed
Test: ACTS test and manual test on DFS channel
Change-Id: I535d360450c874ce05a0b07290326b53578a278b
Signed-off-by: Roger Wang <wangroger@google.com>