bcmdhd: Disabled TCP SYNC flood protection
A connection failure happened with TCP sync flood notification.
The feature was a kind of WAR because old kernel didn't have any protection for TCP SYNC flood attack at that moment.
Currently the kernel has the protection itself with iptables and we doesn't use this feature anymore (from Android Q),
so let's disable this feature.

Bug: 203174019
Test: Validated basic test.
Test: Device is able to connect AP and run speed test.
Change-Id: I8336ff2bafbc2ed08dd33e4d2d1cd1609222d070
Signed-off-by: Roger Wang <wangroger@google.com>