Remove usage of irq_to_desc API in GKI kernels
Analysis & Solution:
Linux has removed it upstream. Instead of the API, Replace it with irq_get_irq_data and irq_data_to_desc.

note: since we are not sure when the irq_to_desc will be deprecated,
so we temporary take 5.11 version at here.
once the version was confirmed, we will modify the version again.

Bug: 203607455
Test: Validated with Slider
Test: Basic function works as expected
Change-Id: Id87f4ce4cc7ca012d4583fe77d737735b1bda874
Signed-off-by: Roger Wang <wangroger@google.com>