bcmdhd: Fixed to add missing p2p_ssid in merge
Original issue & solution:
DHD was filtering SSIDs with "DIRECT" prefix. In corresponding condition, fixed to operate legacy scan instead of goto.
Current issue & solution:
In the process of merging in order to raise the CL, the if statement was changed. The if statement has been modified to its original state.

Bug: 204014524
Test: Validated with Hikey
Change-Id: Ic0f0799fdc74aaacf265dd7741126e2a30b71740
Signed-off-by: chenpaul <chenpaul@google.com>