bcmdhd: Fixed softAP issue for WPA3 R3 certification
WPA3 SAE-SOFTAP 4.2.2.4/4.4.1 TC failed as APUT doesn't advertise "Extended Supported Rates & BSS Membership Selectors element" in its beacon.
When hostapd is configured with sae_pwe = 1, it means H2E support is mandatory for the client to join the AP. This needs to be advertised in RATE/EXT_RATE IE.
We fixed to parse the beacon IEs from hostapd and check for SAE-H2E-BSS-SELECTOR and configure sae_pwe as SAE_PWE_H2E.
The firmware change is required for this fix.

Bug: 203608249
Test: Validated by BRCM cert team
Change-Id: Ib2250c09013eb95d8176269fbd27444a7a2d07b0
Signed-off-by: Roger Wang <wangroger@google.com>