bcmdhd: Fixed to dump pktid map/unmap log properly for debug ring
This is a debug patch to debug dup pktid audit failure issue for GG project.
Currently, pktid map/unmap logging is not available in GG debug_dump.
So, added to dump it properly for GG debug ring.

Since we can't reproduce it until now, we're going to put this change into the formal release.
It would be helpful later. The required FW changes are already included in FW 20.25.

Bug: 211750191
Test: Any issue is not seen during basic sanity & GG aging test
Signed-off-by: Damon Kim <taekhun.kim@broadcom.corp-partner.google.com>
Change-Id: I87b2643f2e3c2337e1bfab825717a1d3b3c6c003