bcmdhd: Fixed to change all RX direction DMAs to bi-directional to prevent S2MPU fault
There are certain Rx data frames which are processed in the dongle (e.g TDLS, Neighbour Discovery(IPV6) offload, etc).
Since these are data packets, they will be received in the host rx buffer by default. As part of processing these frames
in firmware, dongle identifies them, schedules pktfetch & reads the entire host rx buffer to fetch the payload to dongle
for processing them further. Hence the Rx buffers should be marked as bi-directional.

According to the socram, since the Rx buffer address is not marked as bidirectional and
the dongle tried to read this address because of pktfetch, it resulting in S2MPU fault.

This fix already exist for DHD 101.10.531, but, missed it for DHD 101.10.
So, fixed to mark RX DMAs as bi-directioanl properly.

Bug: 215320074
Test: Any issue is not seen during basic sanity
Signed-off-by: Damon Kim <taekhun.kim@broadcom.corp-partner.google.com>
Change-Id: I347a4466320e5b5ba15c73d1f973e04eba0defda