bcmdhd: Fixed to be disabled ipmcast_l2filter if AP operates
Basically, ipmcast_l2filter(multicast filter IOVAR) exist in dongle.
It aimed to reduce host wake by unwanted multicast packets.
It is supposed to be activated only in STA-standalone, but it is enabled even though AP works.

So, when the legacy pixel tries to connect P21/P22 hotspot,
the multicast packets(NS) from the client filtered out,
it's resulting in client disassociation due to no NA packet for the NS packet.

Hence, the driver fixed to be disabled ipmcast_l2filter if AP works.

Bug: 217089990
Test: multicast filter doesn't set properly if AP operates
Signed-off-by: Damon Kim <taekhun.kim@broadcom.corp-partner.google.com>
Change-Id: I2fcbe74456ee38c19458d7b3d6b28d4497bee360