bcmdhd: Fixed Hotspot turn on failure by no available safe channels
So far, sometimes hotspot turn-on failure has been reported,
which is happened due to safe/unsafe channel lists being empty.
After tracking down the root cause, it turned out that the intermittent 'cellavoid' IOVAR failure causes it.

Basically, 'cfg->ioctl_buf' is used in common to send down the IOVAR(wldev_iovar_setbuf_bsscfg),
as this buffer can be used in multiple contexts, it should be protected by the mutex 'cfg->ioctl_buf_sync'.
However, that is missing where 'cellavoid' IOVAR is used. So, it may cause randomly 'ioctl_buf' corruption,
it could be followed by unexpected IOVAR failure.

Hence, 'cellavoid' IOVAR is fixed to be protected with a mutex lock.
And, the debug logs left will be helpful for similar issues later.

Bug: 225576566
Test: Any issue is not seen during basic sanity
Signed-off-by: Damon Kim <taekhun.kim@broadcom.corp-partner.google.com>
Change-Id: I1ca1af584a4da6ba834885eb94068b69d21fa35a