bcmdhd: Added exception handling for SSID length of IE stream
GG reported Kernel Panic issue, which is caused by "system crash 4 times in 10 minutes".
The system crash caused by WifiSsid, it happened due to SSID length(35) exceeded the maximum length(32).

Basically, the host driver have exception handling for SSID length.
However, there are 2 sources, one(ssidie[1]) is come from IE stream, another one(bi->SSID_len) is delivered from FW.

In this case, 'ssidie[1]' indicated 35, 'bi->SSID_len' is 0.
The host driver checks 'bi->SSID_len', but 'ssidie[1]'
DHD checks if the length of 'bi->SSID_len' exceeds the maximum, but not for 'ssidie[1]'.

Hence, in this case, an exception handling is added so that the wrong SSID is not sent up to the upper layer.

Bug: 228164218
Test: Any issue is not seen during basic sanity
Signed-off-by: Damon Kim <taekhun.kim@broadcom.corp-partner.google.com>
Change-Id: Ice64cd4568bc7c799e8d7ab110c1f1928d7f02af