bcmdhd: Added protection logic to avoid unwanted KP at roam scan completion context
KP happened due to out-of-index accessing during roam scan completion.
According to the host ramdump, all of data(roam_log_scan_cmplt_v2_t *) looks corrupted.

There are several possibilities:
- Dongle send-up the wrong information.
- Host driver parsed the event data incorrectly.
- Memory doesn't update properly by the abnormal DMA operation.

However, we have nothing to check more as of now due to a lack of debug info.
So, we assume that FW sent the wrong data to the host(not sure),
and we'd like to add a protection code to avoid KP if it comes again.

Even though the memory corrupted again in the same path, KP will not happen if this protection code is applied.

Bug: 233622771
Test: Any issue is not seen during basic sanity
Signed-off-by: Damon Kim <taekhun.kim@broadcom.corp-partner.google.com>
Change-Id: I040ba05d5c390abf0cbc2b09ca35df1b975fb580