bcmdhd: Fixed the ACS to avoid a restricted channel for SCC case
Analysis:
The country code was set to BE(BELGIUM). The STA was connected to 5G AP that works in Ch44.
Since this channel is a restricted channel, so Softap can't be created.
However, when turning on Softap, the ACS logic picked this channel due to the same channel concurrency.

Root cause:
For SCC case, DHD only checked whether the sta's channel is DFS or not.
So, in this case, ACS picks the sta's channel even though the channel is a restricted channel.

Fix:
When checking the SCC case, DHD also checks if it's a restricted or not.

Bug: 252867295
Test: Verified with Pixel device

Signed-off-by: Dewey Lee <dewey.lee@broadcom.corp-partner.google.com>
Change-Id: If76dab5d91aea795254cb4cc159245819c368ceb