bcmdhd: Reduce the P2P GC assoc scan time
The long P2P connection time is reported.
It shows that 140ms time gap is observed between Join triggered and AUTH complete time.

[13:01:35.915820][  188.329545] wl_conn_debug_info: [p2p-wlan0-1] Connecting with ff:xx:xx:xx:xf:ff ssid_len:11 chan_cnt:1 eidx:1139
[13:01:36.056203][  188.469926] MACEVENT: WLC_E_AUTH, MAC 9a:xx:xx:xx:x9:1a, Open System, SUCCESS

Since the BSSID hint is not used STA use the 80ms assoc scan dwell time with 4 nprobe.

If bssid information is delivered to FW when join is triggered, assoc scan is stopped when it received target GO probe response.
Due to this, Join to AUTH complete time is reduced from 140ms to 50ms.
[13:20:25.442709][  156.889951] wl_conn_debug_info: [p2p-wlan0-0] Connecting with 9a:xx:xx:xx:x9:1a ssid_len:11 chan_cnt:1 eidx:1058
[13:20:25.489951][  156.937190] MACEVENT: WLC_E_AUTH, MAC 9a:xx:xx:xx:x9:1a, Open System, SUCCESS

Bug: 249518141
Test: Verified with Pixel device

Change-Id: Ic20f480eb7ea43931a0b2052690649146f936950
Signed-off-by: Sungjoon Park <sungjoon.park@broadcom.corp-partner.google.com>