bcmdhd: Fixed the ACS 2G SCC case
In b/252867295, we fixed the ACS to avoid a restricted channel for SCC case and modified the “per_chan_info” iovar to be called from one place, so wl_check_active_2g_chan() was removed and was replaced to other functions.
However, the fix for b/252867295 didn't handle 2G SCC case properly when STA was connected to 2G restricted channel like Ch13 on India(IN/0) country.

Thus, fixed the ACS 2G SCC case to get original functionality at the same location.

Bug: 260162012
Test: Verified with Pixel device

Signed-off-by: Dewey Lee <dewey.lee@broadcom.corp-partner.google.com>
Change-Id: I7cab75234e87fb6e5eb1093466e45ca114e20b17