bcmdhd: Avoid connecting when the SoftAP interface is created in progress.
Issue]
This issue happened for below sequence.
1. Start Dual SoftAP.
2. Created interface wlan2/ wlan3.
3. Succeed wlan2 interface up.
4. Connect to AP for STA.
5. Started interface up wlan3 interface.
6. The Link up event did not happen for wlan3, so coredump worked for BY_AP_LINK_FAILURE reason.
We tried to reproduce this issue for use below command during 3days. But couldn't reproduce this symptom.
Maybe it seems that the connection of STA was difficult between the interface up sequence for wlan2 and wlan3.

Fix]
Added a modification for DHD which avoids connection of STA when creating SoftAP.

Bug: 278462235
Test:
This issue could not reproduce. So tested a similar step for 1 day. Below was a tested scenario.
1. Run connecting command(STA).
2. Start SoftAP for 2G and 5G.
3. Wait for 30 seconds.
4. Stop Softap.
5. Disconnect to AP.(STA)
Repeat 1 ~ 5 step.

"AP LINK UP TIMEOUT" log was not observed about 1 day.

Change-Id: I51f1d8ee69896564f5989183b4e1b33a9754de9e
Signed-off-by: Thomson Kang <thomson.kang@broadcom.corp-partner.google.com>