bcmdhd: Added to check overlap channel's channel info for avoiding restrict condition to overlap channel.
[Issue]
Did not removed 165/40 and 165/80 ch for avail channel list in wl_cellavoid_verify_avail_chan_list.
When checking the comment for the comment, wl_cellavoid_verify_avail_chan_list function made purpose for removing the channel.

/* This function is used verifying channel items
 * created by wl_cellavoid_alloc_avail_chan_list are valid
 * by comparing the channel item to chan_info_list from FW
 * If it does not match with chan_info_list, drop
 * CH165 only support 20MHz BW, so 165/40, 165/80 chanspecs created by
 * wl_cellavoid_alloc_avail_chan_list_band are dropped by this function
 */

But did not removed 165/40 and 165/80 caused by this channels was not CLM restrict.

panther:/ # wl chan_info_list
1            (0x1001)2G Band
...snip...
165l         (0xd8a7)5G Band
169u         (0xd9a7)5G Band, CLM Restrict
..snip..
165/80       (0xe0ab)5G Band
169/80       (0xe1ab)5G Band, CLM Restrict

So could select 165/80 channel when detecting lte coex channel.
In this case request to switch 165/80 channel, but failed the CSA command caused by not support the channel in softap.

[Fix]
Check channel info of the overlap channel.
And if the channel was the restrict channel, then removed the channel in avail list.

Bug: 281906332
Test: Tested done for P22.

Change-Id: I51a5c0c529407feb6b147bdd5ff1e13aa47330e3
Signed-off-by: Thomson Kang <thomson.kang@broadcom.corp-partner.google.com>