bcmdhd: Fixed to ignore validations for not supported channels during cellaoid
Issue:
Framework tried to set unsafe channels to the driver, but the driver was setting channel avoidance failed response.

Analysis:
Ch68 is a not-allocated channel.
When the framework sends an unsafe channel list with this channel(ch68) to the driver through cellavoid cmd, the driver returns invalid(-22) and eventually the cmd fails.

Fix:
fixed to ignore validations for not supported channels during cellavoid.

Bug: 293019789
Test: Verified with Pixel device

Signed-off-by: Dewey Lee <dewey.lee@broadcom.corp-partner.google.com>
Change-Id: I067cd5efe358c5a1e3e8ee631831b5f533164130